devnft

+++ install-gui.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 set -e
+export NODE_OPTIONS="--max-old-space-size=3000"
+
 
 if [ -z "$VIRTUAL_ENV" ]; then
   echo "This requires the olive python virtual environment."
@@ -8,7 +10,7 @@ if [ -z "$VIRTUAL_ENV" ]; then
 fi
 
 if [ "$(id -u)" = 0 ]; then
-  echo "The Olive Blockchain GUI can not be installed or run by the root user."
+  echo "The Covid Blockchain GUI can not be installed or run by the root user."
 	exit 1
 fi
 
@@ -23,20 +25,20 @@ if [ "$(uname)" = "Linux" ]; then
 		# Debian/Ubuntu
 		UBUNTU=true
 		sudo apt-get install -y npm nodejs libxss1
-	elif type yum &&  [ ! -f "/etc/redhat-release" ] && [ ! -f "/etc/centos-release" ] && [ ! -f /etc/rocky-release ]; then
+	elif type yum &&  [ ! -f "/etc/redhat-release" ] && [ ! -f "/etc/centos-release" ] && [ ! -f /etc/rocky-release ] && [ ! -f /etc/fedora-release ]; then
 		# AMZN 2
 		echo "Installing on Amazon Linux 2."
 		curl -sL https://rpm.nodesource.com/setup_12.x | sudo bash -
 		sudo yum install -y nodejs
-	elif type yum && [ ! -f /etc/rocky-release ] && [ -f /etc/redhat-release ] || [ -f /etc/centos-release ]; then
+	elif type yum && [ ! -f /etc/rocky-release ] && [ ! -f /etc/fedora-release ] && [ -f /etc/redhat-release ] || [ -f /etc/centos-release ]; then
 		# CentOS or Redhat
 		echo "Installing on CentOS/Redhat."
 		curl -sL https://rpm.nodesource.com/setup_12.x | sudo bash -
 		sudo yum install -y nodejs
-	elif type yum && [ -f /etc/rocky-release ]; then
+	elif type yum && [ -f /etc/rocky-release ] || [ -f /etc/fedora-release ]; then
                 # RockyLinux
-                echo "Installing on RockyLinux"
-                dnf module enable nodejs:12
+                echo "Installing on RockyLinux/Fedora"
+                sudo dnf module enable nodejs:12
                 sudo dnf install -y nodejs
         fi
 
@@ -98,6 +100,6 @@ else
 fi
 
 echo ""
-echo "Olive blockchain install-gui.sh completed."
+echo "Covid blockchain install-gui.sh completed."
 echo ""
 echo "Type 'cd olive-blockchain-gui' and then 'npm run electron &' to start the GUI."
